preliminary layout organization
This time, we conceive of a pericope as the top level, with (parallel) passages containing verses.

Current thinking is to make common/standard objects callouts to html starting with lower-case (e.g., css.html), while actual substitution variables begin with upper case (e.g., Passage).  Then, content that is specific to the thing at hand, in this case pericopes, is inline html and more obvious since it appears next to the common callouts.

I still have to determine how to render each verse in order inside its passage, and each passage in order inside its pericope, and each pericope in order inside the parallel gospels.  Perhaps a hard-coded entry of the pericopes and passages is in order just as with the books of the Bible.....